requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.1.7
django-bootstrap-modal-forms==1.3.1
django-emoji==2.2.2
django-emoji-picker==0.0.4
django-filter==2.1.0
django-widget-tweaks==1.4.3
isort==4.3.16
lazy-object-proxy==1.3.1
logilab-common==1.4.2
logilab-constraint==0.6.0
mccabe==0.6.1
Pillow==5.4.1
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
virtualenv==16.4.3
virtualenvwrapper-win==1.2.5
wrapt==1.11.1
